Add user-specific adjustment of an installation name
Although an installation's name can be configured to control sharing
(especially of packages) among installations, non-privileged users of
an installation could configure sharing only through filesystem links.
An "other-version" directory in the user's add-on directory now serves
as a kind of default link; it applies only when a directory with the
installation's name does not already exist and when user-specific
paths are enabled during Racket's initialization.