rktio: avoid too-wide permissions during file copy
The file-copying operation is supposed to keep the permissions of the
source file, but on Unix, the file was created with default
permissions and then adjusted only at the end --- which might leave
the new copy with termporarily too permissive access. Use the original
file's permissions as an upper bound when creating the copy.