for: add support for `#:splice` clauses
A `#:splice` clause suports interlaving clause binding with clause
expansion in a `for` form (an variants).

The `#:splice` option is somewhat ugly, because `let-values`-style
binding clauses don't naturally accomodate extension. The intent is to
support forms that are built on `for` and that have binding clauses
mixed with expressions in a way that is more like internal-definition
contexts.