rktio: unbreak Windows console input
Commit 081f547a94 incorrect read input as wide-characters without any
translation. This commit restores the same translation that the old
background-thread approach performs.

Both the old and new modes remain present in the implementation,
because there's a trade-off. The new mode works right with expeditor,
because it avoids consuming ready characters in the back group. But
the new mode may claim that input is ready --- and even block Racket
threads --- when there are "uncooked" input events, and Return hasn't
been pressed to make text available. For now, we err on the side of
expeditor.

Related to #4141