Chez Scheme: isolate exact-integers in cptypes lattice
Add an additional field to pred-or to store the information about exact
integer separately from the other number types. This ensures that when a
check (fixnum? x) is false, that information is not lost. In particular,
in
  (unless (fixnum? x)
    (unless (fixnum? x)
      (something x)))
the second check will be eliminated.

Also, this includes fixes that enable some missed reduction
opportunities.