ffi/unsafe: fix potential problem in `cast` from a finalized pointer
This change is meant to repair a failure on DrDr:

 raco test: ".../racket/share/pkgs/draw-test/tests/racket/draw/dc.rkt"

 bytes->string/utf-8: byte string is not a well-formed UTF-8 encoding
   byte string: #"\260H\372P\340U"
   context...:
    .../racket/share/pkgs/draw-lib/racket/draw/private/font.rkt:378:0: get-face-list
    .../racket/share/pkgs/draw-test/tests/racket/draw/dc.rkt:564:0
    body of ".../racket/share/pkgs/draw-test/tests/racket/draw/dc.rkt"
   .../pkgs/compiler-lib/compiler/commands/test.rkt:191:16

A relevant string conversion there is

 (cast (pango_font_description_to_string/ptr desc) _pointer _string)

where the result of `pango_font_description_to_string/ptr` has a
finalizer to free the string. Meanwhile `cast` ends up extracting an
address from the pointer object in the proces sof conversion. It needs
to retain the pointer object meanwhile to make sure no relevant
finalizer runs.