raco setup: fix "info-domain.rktd" update for omitted paths
When `raco setup` checks whether an "info-domain.rktd" needs to be
updated, it tries to drop collections where `compiled-omit-paths` has
a avlue of `'all`. However, the step later that builds
"info-domain.rktd" files did not include that check. So, for a
main-distribution installation, `raco setup` always thought that
"info-domain.rktd" needs to be updated.