Chez Scheme: fix callable floating-opint argument allocation on x86
When a foreign callabale receives a `double` or `float` argument, the
allocation of space to box the number would try to save the
floating-point return register in the (rare) case that allocation
requires a new page of memory. Saving the return register harmless on
most platforms, but on x86, a save and restore involves popping then
pushing the x87 register stack, which is invalid if nothing was there
at the start.