Switch the build system to Zuo
Zuo is a tiny Racket for scripting with facilities inspired by `make`
and Shake. The implementation has a single C file (plus libraries
implemented in Zuo), so it compiles easily. The Racket build
makefiles/scripts now ensure that Zuo is built and then defer most
mork to Zuo scripts.

There's no change to the entry points for building Racket, because
`make` is still supported, and `configure` is still the way to
configure build details. Internally, though, contorted uses of
makefiles, shell scripts, and Visual Studio projects have been
replaced by Zuo scripts that are modular, that work on Windows as well
as Unix, that handle filenames properly (i.e., when paths contain
spaces), and that implement better dependency tracking. For those who
maintain build scripts, the Zuo language should feel like a familar
subset of Racket paired with libraries that implement a make-like DSL.

The directory `racket/src/zuo` is the home of the Zuo implementation,
but much like the Racket branch of Chez Scheme, Zuo will be mirrored
to its own repo. The Racket branch of Chez Scheme is converted here to
Zuo, too.

Incidentally, this commit completes support for 64-bit Arm Windows in
Racket CS and Racket BC.