Use default branch for GitHub package sources
Commit cff766ab84dcc5d9dc969930611e0b8c150bfaf6 changed **Git** package sources
to use the default branch (instead of assuming `master` explicitly), but the
more specialised **GitHub** code path was missed at the time. This fixes the
GitHub path as well.

As with Git package sources at the time of the previous commit above, this is
technically a backward-incompatible change to the interpretation of GitHub
package sources, but explicit branch specification continues to work the same.
For the foreseeable future, to support recent versions, packages in a branch
other than `master` will still need to be specified using the branch name, such
as including `#main` at the end of the package source. Eventually, relevant
versions of Racket will support the new default.

Related to https://github.com/racket/racket/issues/3672
Related to https://github.com/racket/pkg-index/issues/32