bc: repairs for fixnum--flonum comparison
This is the BC counterpart to 6cddc45f68. To restate the problem: on a
64-bit platform, conversion from a fixnum to a flonum can lose
precision, but the implementation of `=`, `<`, etc., did not take that
possibility into account. They converted the fixnum to a flonum for
comparison, but where precision would be lost, the conversion for
comarison needs to go from inexact to exact, instead.