schemify: ensure order of unsafe after potential exception
Schmeify incorrectly conflated "unsafe, so succeeds with the same
value if ever" with "pure, so succeeds at any time". As a result,
it did not convert to force an order in something like

 (list (car x) (unafe-cdr x))

It's difficult to trigger a reordering at the cp0 level for something
like this, and I didn't find a way to do it, but it might be possible
prior to this commit's repair.