Chez Scheme: fix alignment issues for pb
None of the native-code backends currently target an architecture that
requires aligned data access, but pb should be able to run on those
machines. The biggest problem is related to information embedded in
code next to a return address. Add an alignment option for
return-point data, because it's accessed in various ways that would be
inconvenient to make safe for unaligned data. Meanwhile, data literals
in code can be unaligned, still, because those accesses can be made
unaligned-safe relatively easily.