raco pkg: reduce impact of quadratic-time dependency resolution
When `raco pkg install` checks for dependencies and finds new ones,
then it starts over internally with the larger set of packages. This
approach keeps conflict checking simple and consistent, but it also
creates a quadratic-time algorithm. It turns out that there are now
enough packages with documentation for this to matter when the pkg
build service is trying to assemble documentation.

This commit doesn't eliminate all quadratic-time behavior, but the
part that's still quadratic time has a constant factor that is several
orders of magnitude smaller than the part that is no longer quadratic.