CS: separate `prop:equal+hash` from `record-type-equal-procedure`
Although `record-type-equal-procedure` and `record-type-hash-procedure`
were added to Chez Scheme to allow Racket CS to customize `equal?`, it
later turned out that re-impleemnting `equal?` at the Rumble layer was
needed to get the right interaction with impersonators. So,
`record-type-equal-procedure` turned out not to be helpful, and with
the new `prop:equal+hash` protocol to support an equality mode, it
started to get in the way. So, just use the Rumble level's property
support directly, which is simpler and a little faster.

This commit also repairs the interaction of the new `prop:equal+hash`
protocol and chaperones.