add grapheme-cluster functions
The main function is `char-grapheme-cluster-step`, which supports
streaming characters through a counter. The
`string-grapheme-cluster-length` function builds on that to report the
number of characters that continue a grapheme cluster from a given
starting point in a string, and `string-grapheme-cluster-count` builds
on that to count the number of grapheme clusters in a (sub)string.

The `char-extended-pictographic?` and `char-extend?` functions expose
extra character properties that `char-grapheme-cluster-step` uses
internally.