CS: port counting in grapheme clusters instead of characters
Adjust the default port counting function to count in terms of
Unicode grapheme clusters instead of Unicde code points.

The implementation is designed to minimize overhead, and this patch
includes a change to inline `char-grapheme-cluster-step` in unsafe
mode on simple-and-common cases, but the change still significantly
increases the overhead of counting for ASCII input. As a benchmark,
using "expander.rktl" (Racket linklet that turns into "expander.scm"
and is 90k lines, 8.9M bytes) from a file port on AArch64 Mac OS:

           read-byte  read-char  read-bytes read-string  read[*]

 no count     56 ms      73 ms      1 ms      60 ms      352 ms

 old          82 ms      99 ms     31 ms      90 ms      267 ms
 new          94 ms     115 ms     43 ms     102 ms      279 ms

 BC cify     366 ms     403 ms     24 ms      26 ms      495 ms
 BC x86_64   253 ms     297 ms     25 ms      44 ms     1005 ms

   [*] not sure why port counting is *faster* for read;
       read-syntax is a little slower, but about the same

Using 2048 copies of the "hello world" example from the Racket web
page, so that some non-ASCII characters are involved:

           read-byte  read-char  read-bytes read-string  read

 no count     32 ms      47 ms      0 ms      32 ms      392 ms

 old          75 ms      89 ms     23 ms      55 ms      246 ms
 new          88 ms     102 ms     32 ms      64 ms      254 ms

 BC cify     171 ms     204 ms     15 ms      25 ms      473 ms
 BC x86_64   118 ms     142 ms     14 ms      24 ms      968 ms

230,400 pirate flags, each on its own line (so new sees 460,800
positions, while old sees 1,152,000, since a pirate flag is 4 Unicode
code points, and no-counting just sees 3,225,600 bytes):

           read-byte  read-char  read-bytes read-string

 no count     20 ms      41 ms      0 ms      15 ms

 old         128 ms     145 ms     23 ms      39 ms
 new         143 ms     161 ms     30 ms      45 ms

 BC cify     159 ms     196 ms     10 ms      17 ms
 BC x86_64   118 ms     251 ms     10 ms      20 ms