expander: read fewer characters to determine some errors
Commit 4c162c59c0 allows comments produced by a readtable handler in
some places where they should always have been allowed. In that
implementation, in the case of a reader error (i.e., not a comment),
the reader may consume extra characters than before.

This commit narrows the cases where extra characters are read, because
it's often apparent from the first character that a readtable handler
can't be triggered. The character-consuming behavior of the reader for
error cases is then closer to what it used to be for `read`,
`read-syntax` and especially `read-language`.