Windows: work around issue with DLL embedding
Before this change, the current "libgobject-2.0-2.dll" build fails in
embedded-DLL mode. The crash happens during the MinGW-generated pseudo
relocation step at the start of DllMain, because it tries to write to
a page that is write-protected. It's not clear why the relocation
step's attempt to make the page writable doesn't work, but maybe it
has something to do with the way MemoryModule fakes the DLL setup.
Keeping writes enabled until after DllMain returns is enough to avoid
the problem and seems unlikely to create new problems.