raco pkg: add 409 to HTTP results that trigger a retry
It's not obvious that 409 should trigger a retry, but experience with
S3-served catalog information suggests that a retry makes sense. More
generally, a 409 error seems likely to reflect a temporary state.
Also, retries are limited, so there shouldn't be a danger of ignoring
a persistent error or getting stuck.

Closes #4346