Fix #4386
A subclass can share values with its superclass when there are no
changes to the method table. Attaching an interface to a subclass
was not considered a method change, however when that interface
has contracts it will mutable the method table. Sharing in this
case causes the superclass to be messed up. The solution is to
consider interface attachment as a method change so that sharing
no longer occurs.