list: make permutations and in-permutations ordering agree
In `permutations`, we need to reverse the list, because we want to build
the list backward by consing the last element until the first element.
`in-permutations`, however, generates each element from front-to-back,
so we should not need reversing.

Also adjust existing tests so that it checks that both orders agree.