zuo: add support for environment variables to `config-file->hash`
The intent is to support `make`-like behavior where environment
variables are used for things not otherwise configured (by
command-line arguments or a file), but only for environment variables
specifically listed in a `ZUO_CONFIG_ENV_VARS` environment variable.

Requiring variables to be listed in `ZUO_CONFIG_ENV_VARS` is intended
to prevent accidental use of environment variables in configuring a
build.