Chez Scheme: use separate mutex for stdio ports
In a threaded build, Chez Scheme ports are generally not thread-safe,
but the ports for stdin, stdout, and stderr are thread-safe. Those
were previously made thread-safe by taking the tc mutex, which
requires disabling interrupts, which means that Ctl-C doesn't work
when Chez Scheme is blocked on standard input or output. Also, being
blocked on standard input or output means that other threads will get
stuck if they allocate enough to trigger a GC.

The change here uses a separate mutex. To avoid the problem
encountered in an earlier attempt (based on the comments in the code),
the mutex is not taken when the tc mutex is already held (which
can happen when stdio ports are closed on exit, for example).