racket/future: make for{,*}/async return void
The docs already state that the future of touches is not guaranteed
and thus no one should be relying on the result value here.

Additionally, no code[1] on the package server seems to be relying on
the result values from uses of this form, so the likelihood that this
will break anything seems very low.

[1]: https://gist.github.com/Bogdanp/ea71e27ad53339ffb7c0196554b88c88