Chez Scheme: more `configure` and makefile improvements
Fix `configure` handling of an unknown machine type, but infer
compiler flags for a known OS like Linux, etc., on unknown
architecture.

Also, the top-level makefile `boot` and `bootquick` targets were meant
to match the ones before the build-system change, but a different
comvention from the old "Mf-boot" layer was provided. Provide the
other convention as the preferred one.