Support required methods in racket/generic
This PR makes it possible to mark some methods in a generic interface as
"required". Required methods must be implemented in struct's #:methods
clauses. If they are not implemented, a compile time error is raised.

The PR also switches gen:equal+hash, gen:equal-mode+hash,
gen:custom-write, and gen:stream to use this new protocol, mandating
that methods must be implemented. This allows the error message to be
much more better.