Chez Scheme: change `make-record-type-descriptor*` to an overload
Instead of a separate `make-record-type-descriptor*` to create a
record type with anonymous fields, allow a pair of count and mask
in place of the field vector of `make-record-type-descriptor`.

Using a pair doesn't seem as clean as two separate arguments. Although
`make-struct-type-descriptor` could instead distinguish based on the
number of supplied arguments, that would interference with future
additions of optional arguments. Also, extra arguments cannot work
with the internal `$make-record-type-descriptor`, which already
accepts additional arbitrary arguments.

Add `has-` to the name of `record-type-has-named-fields?`. Remove
`record-type-field-indices` and make `record-type-field-names`
report each anonymous field using the name `field`.