search: ranking results based on package status
This is a first cut for search result ranking.
Within a group of search results of the same score,
search results associated with the `base` package will come first,
followed by those in `main-distribution`, and then everything else.

Search results also have an additional indicator on the right border,
which can be hovered to read the description.

Unfortunately, it slows down `raco setup scribblings/main` by about 10
seconds. The command is invoked quite often, as it is run as a part of
`raco pkg install` and `raco pkg remove`.
It is unclear if the performance hit is a deal breaker.

Fixes #4083 and issues raised in
https://racket.discourse.group/t/whats-official-racket-and-whats-not/1447