repair for `thread-wait` and `thread-dead-evt`
When `thread-wait` or the result of `thread-dead-evt` is used to wait
on a thread created with `thread` (as opposed to
`thread/suspend-to-kill`), it may be that the target thread nor the
current thread is otherwise accessible. However, a shutdown action on
a custodian to terminate the target thread would enable the current
thread to continue. Before this commit, the waiting thread might have
been GCed, which isn't correct.

This commit repairs both BS and CS to ensure that the shutdown result
does not get lost. The thread being waited on (not the waiting thread)
might get reclaimed, if possible, but a shutdown action will still
trigger the event that reflects the thread's termination, and so the
waiting thread can continue.

Meanwhile, `thread/suspend-to-kill` needs to interact with
`thread-suspend-evt` and custodians in the same way that `thread` and
`thread-dead-evt` interact with custodians. That's repaired in this
commit, too.

As a special case, `(thread-wait (current-thread))` still allows the
thread (which is both current and the target) to be reclaimed, since
the thread obviously cannot continue after it has terminated. But no
such special treatment applies to waiting on the current thread's
event from `(thread-dead-evt (current-thread))`.

While the old behvior created at least one problem (see
racket/drracket#586), it's possible that it prevented some space leaks
accidentally. We'll have to keep an eye out. If a case is discovered,
a possible solution is to change a `thread` to
`thread/suspend-to-kill`, which creates a thread that is not
terminated by a custodian shutdown, so blocking on that kind of thread
does not prevent GC.