repair winder sharing detection for non-composable continuation jumps
Score another win for random testing via a Redex model.

The implementation of dynamic-wind frame sharing (to determine which
ones to unwind and wind on a jump) was not consistent with the model
in the ICFP'07 paper. The implementation detected sharing only when a
combination of winder and prompt frames match, while the paper's model
matches only winders independent of prompt frames.

One sign that the paper is the right choice, as opposed to the
implementation: When a composable continuation is applied and not in
tail position with respect to an existing prompt, then an extra
"compose" prompt is inserted in the CS implementation. That prompt is
supposed to be invisible, and special care is take to make sure that
its' not detectable through continuation marks, but winder matching
was affected.

Otherwise, BC and CS were previously consistent with each other, but
not the paper, when an extra prompt is present. For example, when
jumping from

   prompt 1
     |
    ...A ; some frames without DW
     |
    DW_1
     |
    ...B ; some other frame without DW
     |
    [ ]

to

   prompt 1
     |
    ...C
     |
   prompt 2
     |
    ...D
     |
    DW_1
     |
    ...E
     |
    [ ]

should DW_1 be unwound and wound again? The paper says no, while the
implementations said "yes". A "yes" is plausible, because maybe prompt
2 is something like an exception handler and you want to re-run the
DW_1 entry under that handler. But a "no" is also plausible, because
maybe it reflects acquisition of a resource that you're going to keep
using, even if it's in a new context that happens to include something
that you didn't know is implemented with a prompt.

The paper's model (i.e., "no") is simpler to specify and seems more
composable, so the changes here bring the implementations in line.

Closes #4515