extend `copy-file` with additional permission-management arguments
Support a `#:permissions` optional argument as well as one that
control the interaction with a umask when the destination file is
created.

The optional by-position `exists-ok?` argument is still supported, but
a new `#:exists-ok?` keyword argument is also available; a contract
exception is raised if both the by-position and keyword variants are
supplied.

Related to #4511