Chez Scheme: avoid `list?` crash if argument is mutated
This is the solution described in cisco/ChezScheme#599.

One extra test in the `list?` implementation is cheap, since it's a
highly predictable branch. The effect of that check is maybe less
important, since the behavior of list operations just has to be
undefined when the list is mutated concurrent to the operation. It's
nice to raise an exception when undefined behavior is detected in safe
mode, but sometimes that's too expensive; then, any convenient
behavior can be ok, but avoiding memory faults is a priority.