CS: update and expose process-executable finder
Use improved code to find the current executable on some platforms
from LLVM, and expose Racket's implementation to embedding
applications through a new `racket_get_self_exe_path` function.

The implementation ends up duplicated in compiled form, since it needs
to be exported from the Racket shared library on some platforms, while
it is also needed in the main executable potentially to load the
shared library. But the code is small enough that duplication
shouldn't matter.

Closes #4483