Chez Scheme: hashtable cell immutability checks cleanup
Say which ops returning hashtable cells now have immutability checks,
and add missing mutability check to hashtable-ref-cell symbol ht branch.

Now consistently checks for hashtable mutability before checking that the
given key is a symbol for symbol hashtable operations.  That allows the
mutability checks in the generic dispatch versions of hashtable-cell and
hashtable-ref-cell to be placed before the type dispatch.