racket/unit: refactor many forms to use syntax/parse
In addition to altering many of the core unit forms to use syntax/parse,
this commit also refactors the core signature import/export processing
logic to use the `signature` struct rather than passing around nested
lists of a particular shape. This makes much of the code significantly
easier to understand.