Revert "search: ranking results based on package status"
Ranking results is a good idea, but it's not clear that this is the
right approach, and we have another idea to try. Meanwhile, this
initial attempt was accidentally merged (my fault!). We decided to
keep it for a while to see how well it plays, but it turns out that
there is still a problem handling `planet` module paths. Sorawee and I
thikn it's safest to revert for now and try again later.

This reverts commits
 b4b16eccc9299da11c83a341d55500ae3ef49f3f.
 97c76e98625021c84a9bc5c4cd868c8b52c1c37a.
 3e56962e7347b38090cc7c68e149081648b6dd7d.
 bef9f78279d689b14bbae9b61ac637e77bd707b3.
 38638885b58b062520be67859387a50eba0cb744.
 6391481c07935d49ae529d0c27979ae0c3a296fa.
 2f76f580da7f71d57d421516b7a3eadd96f729bb.

(cherry picked from commit fd14868d4a7966b467fa399ef5b357ca7a13ca54)