Support multiple values in stream-cons
The stream protocol and stream operations support multiple values.
However, stream-cons does not. The consequences are:

- Multiple values in streams can't be easily used by users.
- Internal stream operations are implemented in a convoluted way
  to support multiple values, which is error-prone.

  There are many mistakes in the current implementation,
  including but not limited to:

  - Incorrect laziness
  - Linear time memory consumption on operations that should take
    constant time

  See
  https://racket.discourse.group/t/stream-filter-not-in-constant-space/1643
  for the discussion.

This PR adds a support of multiple values to stream-cons.
We then utilize this stream-cons on the stream operations
to fix the issues mentioned above.

Note that there are still known deficiencies in the current
implementation. These deficiencies are not degradation, however, because
SRFI-41 also has them. Furthermore, the data definition of streams in
Racket makes it difficult to fix these deficiencies without a large
refactoring, so I would like to postpone fixing these issues
to the next PRs.

See https://srfi-email.schemers.org/srfi-41/msg/21844204/
for the discussion.