CS: avoid excessive external-event checking
Commit cbbd3c262a included a backward test for whether to propagate
remaining ticks to the Racket thread scheduler, which resulted in
excessive and unproductive polling for system events.

An existing test manages to catch this scheduling inefficiency. But
because it's an unreliable timing test, it took a while for the
combination of a failure happening consistently enough plus sufficient
investigation to determine that things really had been better before.