expander: speed up module-tree instantiation
When a module and all of its dependencies are instantiated for
cmopilation or attached to a new namespace, all of the transitive
requires are checked. Naturally, many paths lead to the same module,
and instantiation avoids traversing already-instantiated subtrees.
Still, there's some overhead in determining a module's resolution and
discovering that it has been visited already.

The expander now keeps track of which immediate `require`s are to
modules that are also required by other `require`s, and instantiation
traverses only the ones that aren't. Often, this compile-time
simplification cuts in half the number of `require`s that are
traversed. For example `(make-base-namespace)` is almost twice as
fast, since most of its time is just traversing the modules that
implement `racket/base`.

The optimization relies on a corerspondence between compile-time and
run-time module resolution that an uncooperative module resolver could
defeat, but the assumption is in line with others made by the compiler
(to enable corss-module inlining, for example).