fix `enter!` and `dynamic-rerequire` for `#lang` search fallback
Changes in 6d1cb97b37 affected whether `get-module-code` raises
`exn:get-module-code` when a module does not exist, and
`dynamic-rerequire` (which is used by `enter!`) was relying on
catching that exception to implement a quiet-failure mode for
submodule loading, which is in turn needed for the `#lang` search
process that tries a `reader` submodule before a "lang/reader.rkt"
relative module.

We could change `get-module-code` again, but it's better to avoid
catching exceptions or otherwise communicate through dynamic scope,
which always risks catching too much. The revised strategy here takes
advantage of the `notify` argument to `get-module-code` as a hook to
check for a file's existence.