add `default-global-port-print-handler`
This handler is the initial value of `global-port-print-handler`, but
it was inaccessible after the parameer is changed --- on purpose, to
discourage resetting the print handler away from one that an
environment (like DrRacket) wants a program to use. But the default
handler doesn't do anything privileged, and it's occasionally useful
to be able to reach the default behavior.