add support for a `configure-expand` submodule
A `configure-expand` submodule configures parameters to customize
dynamic-at-compile-time properties of the environment. Typically,
they're the same parameters that make sense to customize in
`configure-runtime`, but only one `configure-runtime` is used on
Racket startup, while `configure-expand` needs to be used for each
module that is expanded. So, instead of a side effect,
`configure-expand` provides `enter-parameterization` and
`exit-parameterization`, where the former returns a parameterization
to wrap the module's expansion, while the latter returns a
parameterization to wrap the expansion of modules that happen during
the entered expansion.