match: make duplicate hash keys always fail
In the hash-table pattern, the ground truth is the generic case, which
compiles the pattern to list-no-order. Hence, patterns with duplicate
keys will always fail.

The PR #4532 has a mistake because it allows duplicate keys with ..0
to match successfully. This PR amends #4532 to always fail such cases.

Also added a history note that should have been added in #4532