Recommend 100 columns instead of 102 columns
The current recommended line width in the Racket style guide is [102 columns](https://docs.racket-lang.org/style/Textual_Matters.html#%28part._.Line_.Width%29). This is a rather bizarre number; more traditional line widths would be 80, 100, or 120 columns. I can't find any information about where the extra 2 columns come from or what their purpose is. This pull request changes the recommended line width to 100 columns to make it less bizarre.

If there's a reason it's 102 columns instead of 100 and we want to keep it at 102 columns, let me know so I can instead open a pull request to update the docs with the rationale. It's a strange enough choice that I think some explanation would be helpful.

**Changing this line width to 80 or 120 columns is explicitly out of scope for this pull request and discussion.** I'm just trying to tweak the line width to something normal, not start a giant bikeshedding session.