Chez Scheme: release notes: clarify typical effect of cptypes
The cptypes pass was intended to improve run-time performance, and it
somtimes helps a lot, but measurements on realistic programs suggest
that run-time improvements are typically modest. Furtunately, compile
time tends to benefit from the extra pass, even when run time doesn't,
so it's an easier call to have it enabled by default.