Chez Scheme: constant top-link offset for rp header compact and not
Put the `top-link` field the same distance from the end of a compact
rp header and a non-compact rp header. That way, the GC can avoid an
indirection finding code as it traverses a continuation. This can save
collection time on some x86_64 (such as an Intel MacBook) for a
program that captures and saves many continuations, like the one
below.

(define (go)
  (define (make-one)
    (call/cc
     (lambda (esc)
       (let loop ([i 1000])
         (if (zero? i)
             (call/cc esc)
             (+ i (loop (sub1 i))))))))

  (define l
    (let loop ([i 1000])
      (if (zero? i)
          '()
          (cons (make-one)
                (loop (sub1 i))))))

  ;; quadratic-time traversal that creates continuations
  ;; with a spectrum of lifetimes
  (let loop ([l l])
    (unless (null? l)
      ;; replace all in this `l`
      (let loop ([l l])
        (unless (null? l)
          (set-car! l (make-one))
          (loop (cdr l))))
      ;; recur with the rest of `l`
      (loop (cdr l))))

  ;; keep the original list, so we retain the
  ;; continuatons
  l)

(time (begin (go) (void)))