Modify `ssl-load-private-key!` to accept a `bytes` containing the key in PEM format, in place of a pathname
Currently in order to load a private key into an SSL context, it must be written to disk. For applications that have the key in memory, they must currently write it out to a temporary file, which is undesirable for such sensitive data.

This extends ssl-load-private-key to load the key directly, making use of openssl's `PEM_read_bio_PrivateKey` functions to parse and then load the key in to the context.