add `immutable-string?`, etc.
Compensate for the old mistake of `immutable?` by adding
`immutable-string?`, `mutable-string?`, etc., to a `racket/mutability`
library.

Since packages have (very reasonably) defined functions like
`immutable-string?` using the same name, these new functions are not
provided by `racket/base`. So, using one of the functions is not as
convenient as it should be, but at least the primitive implementations
of the predicates are now accessible.