Correct docs for `write-resources`
In testing, I have not been able to get
write-resource to return #f when writing to the 
registry; it always falls back to writing to 
mred.ini and returns #t. This edit brings the
docs in line with that reality. 

Arguably there should be a way to force various
kinds of exceptions if the registry write attempt
fails (due to permissions or to non-existent keys
with create-key? set to #f).