expander: fix `module->namespace` for space definitions
When a module contains an identifier defined in a space (i.e., with an
interned scope on its identifier), then `module->namespace` needs to
remember the binding instead of trying to reconstruct it from a table
of definitions, even if there's no literal syntax object within the
module (which would otherwise be the only route to accessing bindings
that have extra scopes).

Thanks to @Gopiandcode for the bug report.