Windows: `file-or-directory-modify-seconds` write for dirs
Use `SetFileTime` instead of `utime`, because `utime` doesn't
work for directories.

Another effect of avoiding `utime` is that we don't need to try
to be consistent with it when getting a timestamp, which involves
a questionable correction for the case that DST is in effect now
but not at the file's time. That is, the questionable correction
is now disabled. This can change the timestamps that Racket reports
relative to old Racket versions, but a key property is preserved:
after writing a timestamp, reading the timestamp will return the
written value.

Closes #4636