Fix @history from previous commit
I didn't realize that "#:changed" must be used in order to include a
content expression. Anyway "changed" is as good a verb as "added"
for this.

Interestingly the checks run for pull requests on GitHub didn't catch
this problem; maybe those don't build docs? I only noticed it from
building locally.