DRAFT RFC enhancement to provide transformer `export` struct
This does not compile. For example we can't actually import/use
format-id from racket/syntax here.

The point of this PR is to outline a change where the old `out-sym`
field of `export` is now `out-sym-or-id`, and, although an identifier
is preferred, a symbol is still supported for backward compatibility.