fix recursive-contract when it is used in a way that'll get different blame
closes #4648

with the previous implementation, the memoization was closing over a
possibly-stale blame object. In this version, instead of just blindly
using the cached function (which might have the stale object), we
check to see if the negative and positive fields are the same before
we use it. We actually keep two entries in the cache: one when the
blame is the same and one when the blame is reversed, just in case
there is some tricky loop somewhere that would benefit from the
memoization but goes through a negative position (I don't think this
can happen in TR because one side of all contracts never blames anyone
but am not 100% sure).

The cache doesn't use the rest of the blame object because it contains
information that actually does change (the information that feeds into
the "in: ..." portion of the contract error message) but we're okay to
forget how many times we've gone around the recursive contract