Use property value recommended by rfindler
  <https://github.com/racket/racket/pull/4649#issuecomment-1576770816>

Albeit with the change I described here:

  <https://github.com/racket/racket/pull/4649#issuecomment-1579081347>

But that would be easy to change back if desired.

When I update pdb to synchronized with the new property style, all
tests pass. (Note that pdb is effectively still assuming that the
property is attached to the identifier that it describes, because
then (in that low-level `rename` clause in which the id appears) is
when I want to use the value. So for now I am not jumping through any
hoops with side hash-tables or thunks, but could do so someday if the
attach point changes.)