Update docs for export-out-id and import-or-export-prefix-ranges
The convention seems to be that details of the value of the
import-or-export-prefix-ranges syntax property belong in the drracket
repo, in tools.scrbl. So this commit doesn't go into that; it just
mentions that the property is added for prefix-{in out}.

(However it would be weird to document the stx prop value there,
today, because that page is about things check-syntax looks for -- but
it doesn't yet look for this property. So for now the gist of the docs
lives in comments in reqprov.rkt. I plan to update tools.scrbl when I
update check-syntax to look for the prop and call some new
"add-import-or-prefix-ranges" syncheck method.)